Refactored the .gitlab-ci.yml file to reduce redundancy and improve the workflow.
This change introduces a .install-dependencies stage to the CI configuration, which removes the repetitive task of specifying the same 'image' and 'composer install' for each job. It also updates the phpstan and phpunit from being a dev-dependency in the composer.json file to be installed via phive, providing version isolation and ensuring that the tools will not conflict with the project's dependencies. Furthermore, the Laravel pint package was updated to version v1.13.6. This change was made to streamline the build process and keep dependencies up-to-date.