Update .gitignore and composer.lock files
In our .gitignore, we've simplified the file by removing the listing of each individual asset in the /_site and /tools directories for ignoring.

Additionally, we've updated the composer.lock file to reference the newer versions of our packages, including laminas/laminas-hydrator and laravel/pint. This will ensure that we are using the most up-to-date and secure versions of these dependencies."