Removed redundant 'composer install' calls in GitLab CI
In this change, redundant 'composer install' commands were removed from various stages (lint and test) in the .gitlab-ci.yml file. Since 'composer install' is already called in the '.install-dependencies' template, it does not need to be called again in the individual stages, making the CI process cleaner and more efficient.